FIX: refactor data model for reactions (#3)
* FIX: refactor data model for reactions

This was inspired by Sam during our Friday call. He said that because that plugin may become very popular we need to be sure it is performant. He suggested that we should cache count, something similar to `likes_count` column on posts. I was thinking over the weekend and that gave me the idea that we should restructure it a little bit.

I think we should have two tables in the database.

One is `reactions` which would contain information about reactions and cache count:
```
      t.integer :post_id
      t.integer :reaction_type
      t.string :reaction_value
      t.integer :count_cache, default: 0
      t.timestamps
```
and `reaction_users` which would give us information who reacted:
```
      t.integer :reaction_id
      t.integer :user_id
      t.timestamps
```

That simplified logic around serializer which I was worried about. In theory, it was a simple ruby calculation but for a very popular topic with many reactions, it might be a bottleneck.
https://github.com/discourse/discourse-reactions/blob/master/plugin.rb#L52