FIX: Post expanding in activity stream was causing reload (#115)
When clicking on the expand arrow in the reactions activity
stream, a full page reload occurred because core calls:

```
const postNumber = item.get("post_number");
```

Where item is reaction.post. However, reaction.post had not been
converted to an Ember object via the Post model, so .get caused
an error, which triggered the reload. This commit fixes the issue
by making sure that reaction.post is an instance of Post model,
and also fixes an error in the discourse-reactions-reaction-post
template not showing the correct expanded content when the activity
is expanded.