FIX: Include likes in the reactions received list. (#131)
We no longer use like notifications to reduce the overlap between them and reactions. Instead, we replaced them with reaction notifications that display the heart icon, looking the same. Since we replaced them, clicking a consolidated reaction notification will point users to the reactions received page, which doesn't list likes, so the user won't seif one of the consolidated was a "like" the user won't see it.

To fix this, we'll create a reaction user object alongside the post-action when toggling a like, including it in the list. Additionally, if all the consolidated reactions were likes, the notification links to the likes received list instead.