FIX: Remove thumbsup reactions when there's also a +1 reaction. (#141)
The thumbsup reaction didn't work, so we recommended using +1 instead as a workaround. The post will have both reactions associated if someone reacted with a thumbsup, then it was renamed to +1, and another user reacted with a +1.

The post-migration to rename thumbsup reactions we added in a previous commit won't work in the above scenario, so we should directly remove them and leave the existing +1 reactions.