FIX: prevents triggering a reaction while scrolling
It's a simple hack to mark a current touch sequence touchstart/touchend as invalid if the user has been moving finger (scrolling most likely) during start and end.

Not it will require https://github.com/discourse/discourse/pull/22635 to be merged for this to work, but it will just do nothing and not error if the commit is not present as `touchMove` won't be called in this case.