FIX: update post id for reactions when post is moved (#250)
Previously all emoji reactions on a post would be lost when posts are moved to a new topic. This would only happen for the first post in a topic. This is due to an entirely new post being created based on the attributes of the original post.

The reaction data for the old post remains but the new post does not have any.

This change creates a copy of the reaction data and user reaction details from the original post (with updated post id).