FEATURE: Count all reactions as likes with exceptions controlled by a site setting (#267)
This commit changes the reaction paradigm to instead count all reactions as
Likes (via a `PostAction` record) rather than only considering a single specific
reaction emoji as a Like equivalent.

A new site setting, `discourse_reactions_excluded_from_like`, is introduced defaulting
only to have :-1: . Any emojis on this list will **not** create a `PostAction` record that
represents a Like.

We now create a `PostAction` record along with the existing `ReactionUser` record
for reactions which are not excluded from Like, to help with history correction if
`discourse_reactions_excluded_from_like` changes.

For existing settings:

* `discourse_reactions_reaction_for_like` is still used, since we need an easy way to map
  regular Likes to reactions if the plugin is enabled/disabled. This cannot be set to anything
  in `discourse_reactions_excluded_from_like`. If this is changed, history will not get changed.
  Any reaction using this setting will only make a `PostAction` record.
* `discourse_reactions_like_icon` is still used, since that is purely a visual setting.

Also included a job that performs data consistency operations whenever
`discourse_reactions_excluded_from_like` is changed, to create or trash
relevant `PostAction` records, and recover already trashed ones as needed.
Also `UserAction` records for `LIKE` and `WAS_LIKED` are created and deleted
as needed.

For now `discourse_reactions_like_sync_enabled` controls whether we create/trash
`PostAction` and `UserAction` records when the `discourse_reactions_excluded_from_like`
is changed. We are not including a migration until this is running in the wild for a few weeks.