FIX: Hovering reaction icon causing error would flood AJAX requests
If a user hovered over a reaction to see a list of users,
mostly a problem when anon browsing (see
https://github.com/discourse/discourse/commit/5935148bd863ea80865cee1049b3a44f3fbf3cc8),
we didn't have a throttle when erroring, so extra hovers would cause
error messages.

This fixes the issue by adding a throttle, and also adds some general
reactions system testing for this area.