FIX: Do not show users who Reacted under post ... menu
There is a "show-more-actions" button represented by ...
at the bottom of each post. This shows the users who have
liked the post. However, now that reactions also count as likes,
we don't want to show the users who reacted to the post here,
only the ones who liked it with main_reaction_id.

We can use the plugin modifier introduced in core in
https://github.com/discourse/discourse/pull/25740 for this.